Don't store tasks in XmlTaskFactory, just emit them
The tasks are still created on the heap and are sorta the responsibility
of the caller, but now they get the factory as a default parent,
so if nobody takes them the factory will free them as it dies. Otherwise
they will probably end up in some container that will take ownership
and handle them later.