First implementation of getFrob
Preparing for actual auth API usage. The main layout fires off the API
object initialization, which in turn sends the necessary network request
to get a frob for auth. Access to the network is controlled with a
semaphore, since we're supposed to throttle back to a maximum of
one request/sec anyway.

There's a problem with the signing as is however.